refactor: Split end_to_end.rs into focused test modules
Split the large end_to_end.rs test file (1,437 lines) into six smaller,
focused test modules organized by feature area:

1. e2e_basic_crud.rs (175 lines) - Basic SELECT, WHERE, column selection
2. e2e_data_types.rs (590 lines) - Numeric types, CHAR, DISTINCT, GROUP BY, LIMIT
3. e2e_predicates.rs (290 lines) - LIKE, IN, EXISTS predicates
4. e2e_functions.rs (170 lines) - COALESCE, NULLIF functions
5. e2e_complex_queries.rs (210 lines) - Quantified comparisons (ANY/ALL/SOME)
6. e2e_set_operations.rs (230 lines) - UNION, INTERSECT, EXCEPT operations

All tests preserved and passing. Each file is < 600 lines for easier navigation.
